+++ cmd/plumb/report.go
@@ -32,7 +32,7 @@ Examples:
 
 	open, out, title := addReportFlags(fs)
 
-	if err := fs.Parse(args); err != nil {
+	if err := fs.Parse(reorderArgs(fs, args)); err != nil {
 		return err
 	}